# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca Pública",
        "description": "Documentación de la API REST para gestión de autores y libros.",
        "contact": {
            "email": "soporte@biblioteca.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registro de nuevos usuarios",
                "operationId": "0b5e98db9844fcec146cc2029ff2640b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario registrado correctamente"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/v1/authors/{author}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Mostrar detalles de un autor",
                "operationId": "b0f2baefd5fe0363db99a0f93a104f1f",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "ID del autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del autor obtenidos correctamente"
                    },
                    "404": {
                        "description": "Autor no encontrado"
                    }
                }
            }
        },
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Listado de todos los libros",
                "operationId": "ab6891af84b2f0a68a33cdfe8324ce98",
                "responses": {
                    "200": {
                        "description": "Listado obtenido correctamente"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registro de nuevo usuario (Bibliotecario o Directivo)",
                "operationId": "6cb93c5db517a3e6e369c44ede7f22fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan.perez@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Registro exitoso, redirige al login de Filament",
                        "headers": {
                            "Location": {
                                "description": "Redirección al login",
                                "schema": {
                                    "type": "string",
                                    "example": "/admin/login"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (faltan datos o formatos incorrectos)"
                    }
                }
            }
        },
        "/libros": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "summary": "Listado de todos los libros disponibles",
                "operationId": "6a23ed4b99b1c7bfc8d71eda3e67b8b1",
                "responses": {
                    "200": {
                        "description": "Lista de libros obtenida exitosamente."
                    }
                }
            }
        },
        "/autores/{author}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "summary": "Muestra los detalles de un autor específico",
                "operationId": "4ee6ea217149119a6c5807fa89618a6d",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "ID del autor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del autor obtenidos exitosamente."
                    },
                    "404": {
                        "description": "Autor no encontrado."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Author": {
                "title": "Autor",
                "description": "Información de un autor",
                "properties": {
                    "id": {
                        "description": "ID del autor",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre completo del autor",
                        "type": "string"
                    },
                    "birthdate": {
                        "description": "Fecha de nacimiento del autor",
                        "type": "string",
                        "format": "date"
                    },
                    "nationality": {
                        "description": "Nacionalidad del autor",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación del registro",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Book": {
                "title": "Libro",
                "required": [
                    "id",
                    "title",
                    "isbn",
                    "publication_year"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "publication_year": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "Usuario",
                "description": "Información de un usuario registrado",
                "properties": {
                    "id": {
                        "description": "ID del usuario",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nombre completo del usuario",
                        "type": "string"
                    },
                    "email": {
                        "description": "Dirección de correo electrónico del usuario",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Fecha de verificación del correo electrónico",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación del usuario",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización del usuario",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Libros",
            "description": "Operaciones relacionadas con los libros"
        },
        {
            "name": "Autores",
            "description": "Operaciones relacionadas con los autores"
        },
        {
            "name": "Usuarios",
            "description": "Usuarios"
        }
    ]
}